split easygopigo3 into 2 files: easygopigo3 and easy_sensors
